% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{appointments_medication}
\alias{appointments_medication}
\title{patient appointment list combined with medications}
\usage{
appointments_medication(
  dMeasureMedication_obj,
  n_medication = NA,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{dMeasureMedication_obj}{R6 object}

\item{n_medication}{minimum number of medications}

\item{screentag}{if neither \code{screentag} or \code{screentag_print} is
defined then values are derived from \code{self$printcopy_view}}

\item{screentag_print}{}
}
\value{
dataframe of apppointments
$Patient, $AppointmentDate, $AppointmentTime,
$Provider, $Status, $Label
}
\description{
derived from dM$appointments_filtered_time
}
