% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Medication module - server}
\usage{
datatableServer(input, output, session, dMMedication)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMMedication}{dMeasureMedication R6 object}
}
\value{
none
}
\description{
Medication module - server
}
