% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medication.R
\name{dMeasureMedication}
\alias{dMeasureMedication}
\title{dMeasureMedication}
\description{
dMeasureMedication - medication module for dMeasure

list appointments (with medication information) by clinician providers
}
\details{
dMeasureMedication class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dMeasureMedication$new()}}
\item \href{#method-appointments_medication}{\code{dMeasureMedication$appointments_medication()}}
\item \href{#method-clone}{\code{dMeasureMedication$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureMedication$new(dMeasure_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-appointments_medication"></a>}}
\if{latex}{\out{\hypertarget{method-appointments_medication}{}}}
\subsection{Method \code{appointments_medication()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureMedication$appointments_medication(
  n_medication = NA,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureMedication$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
