% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Medication module - server}
\usage{
datatableServer(id, dMMedication)
}
\arguments{
\item{id}{id}

\item{dMMedication}{dMeasureMedication R6 object}
}
\value{
none
}
\description{
Medication module - server
}
